Add recurrence tests and refactor test setup
